Refactor some packages Update Readme
