+++ composeApp/src/androidMain/kotlin/com/macaosoftware/component/navigationcompose/demo/MainActivity.kt
@@ -9,10 +9,10 @@ import androidx.compose.material3.MaterialTheme
 import com.macaosoftware.app.MacaoApplication
 import com.macaosoftware.app.MacaoApplicationState
 import com.macaosoftware.app.startup.task.StartupTaskRunnerDefault
-import com.macaosoftware.component.navigationcompose.demo.startup.DemoRootGraphInitializer
-import com.macaosoftware.component.navigationcompose.demo.startup.DatabaseMigrationStartupTask
-import com.macaosoftware.component.navigationcompose.demo.startup.LaunchDarklyStartupTask
-import com.macaosoftware.component.navigationcompose.demo.startup.FirebaseConfigStartupTask
+import com.macaosoftware.component.navigationcompose.demo.startup.initializers.DemoRootGraphInitializer
+import com.macaosoftware.component.navigationcompose.demo.startup.tasks.DatabaseMigrationStartupTask
+import com.macaosoftware.component.navigationcompose.demo.startup.tasks.LaunchDarklyStartupTask
+import com.macaosoftware.component.navigationcompose.demo.startup.tasks.FirebaseConfigStartupTask
 
 class MainActivity : ComponentActivity() {
