+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/di/ServerUiModule.kt
@@ -1,7 +1,7 @@
 package com.macaosoftware.component.navigationcompose.demo.di
 
 import com.macaosoftware.component.core.DestinationRendersRegistry
-import com.macaosoftware.component.navigationcompose.demo.serverui.DemoDestinationRendersRegistry
+import com.macaosoftware.component.navigationcompose.demo.marketplace.DemoDestinationRendersRegistry
 import com.macaosoftware.component.navigationcompose.demo.serverui.domain.ServerJsonManager
 import com.macaosoftware.component.navigationcompose.demo.serverui.domain.ServerUiDestinationInfoParser
 import org.koin.core.module.dsl.factoryOf
