+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/startup/tasks/DatabaseMigrationStartupTask.kt
@@ -1,13 +1,14 @@
-package com.macaosoftware.component.navigationcompose.demo.startup
+package com.macaosoftware.component.navigationcompose.demo.startup.tasks
 
 import com.macaosoftware.app.startup.task.StartupTask
 import com.macaosoftware.util.MacaoResult
 import kotlinx.coroutines.delay
 import org.koin.core.component.KoinComponent
 
-class LaunchDarklyStartupTask : StartupTask {
+class DatabaseMigrationStartupTask : StartupTask {
+
     override fun name(): String {
-        return "LaunchDarkly remote syncing ..."
+        return "Database Migration"
     }
 
     override fun shouldShowLoader(): Boolean {
