+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/startup/tasks/FirebaseConfigStartupTask.kt
@@ -1,14 +1,14 @@
-package com.macaosoftware.component.navigationcompose.demo.startup
+package com.macaosoftware.component.navigationcompose.demo.startup.tasks
 
 import com.macaosoftware.app.startup.task.StartupTask
 import com.macaosoftware.util.MacaoResult
 import kotlinx.coroutines.delay
 import org.koin.core.component.KoinComponent
 
-class DatabaseMigrationStartupTask : StartupTask {
+class FirebaseConfigStartupTask : StartupTask {
 
     override fun name(): String {
-        return "Database Migration"
+        return "Fetching initial app.config in firebase"
     }
 
     override fun shouldShowLoader(): Boolean {
