+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/startup/tasks/LaunchDarklyStartupTask.kt
@@ -1,14 +1,13 @@
-package com.macaosoftware.component.navigationcompose.demo.startup
+package com.macaosoftware.component.navigationcompose.demo.startup.tasks
 
 import com.macaosoftware.app.startup.task.StartupTask
 import com.macaosoftware.util.MacaoResult
 import kotlinx.coroutines.delay
 import org.koin.core.component.KoinComponent
 
-class FirebaseConfigStartupTask : StartupTask {
-
+class LaunchDarklyStartupTask : StartupTask {
     override fun name(): String {
-        return "Fetching initial app.config in firebase"
+        return "LaunchDarkly remote syncing ..."
     }
 
     override fun shouldShowLoader(): Boolean {
