+++ composeApp/src/jsMain/kotlin/com.macaosoftware.component.navigationcompose.demo/Main.kt
@@ -5,10 +5,10 @@ import androidx.compose.ui.window.CanvasBasedWindow
 import com.macaosoftware.app.MacaoApplication
 import com.macaosoftware.app.MacaoApplicationState
 import com.macaosoftware.app.startup.task.StartupTaskRunnerDefault
-import com.macaosoftware.component.navigationcompose.demo.startup.DemoRootGraphInitializer
-import com.macaosoftware.component.navigationcompose.demo.startup.DatabaseMigrationStartupTask
-import com.macaosoftware.component.navigationcompose.demo.startup.LaunchDarklyStartupTask
-import com.macaosoftware.component.navigationcompose.demo.startup.FirebaseConfigStartupTask
+import com.macaosoftware.component.navigationcompose.demo.startup.initializers.DemoRootGraphInitializer
+import com.macaosoftware.component.navigationcompose.demo.startup.tasks.DatabaseMigrationStartupTask
+import com.macaosoftware.component.navigationcompose.demo.startup.tasks.LaunchDarklyStartupTask
+import com.macaosoftware.component.navigationcompose.demo.startup.tasks.FirebaseConfigStartupTask
 import org.jetbrains.skiko.wasm.onWasmReady
 
 fun main() {
