--- macao-navigation-compose-app/README.md
@@ -1,3 +1,0 @@
-A full compose application architecture package with Koin as dependency injection
-container. It also has an application class with delegates to set the desired plugins.
-implementation("io.github.pablichjenkov:macao-sdk-koin:0.6.10-rc02")
